- Rewrote vec3 & vec2 - Fix placement problem on EXP_GUI_ColorCache

+++ src/ExpansionGameEngineCore/EXP_GUI_ColorCache.cpp
@@ -11,15 +11,15 @@ EXP_GUI_ColorCache::EXP_GUI_ColorCache(EXP_Game* game, vec3f color, float opacit
 
 	RD_GUI_Manager* m_manager = game->GetRenderer()->GetGUI_Manager();
 
-	int w = game->GetRenderer()->getWindowWidth();
-	int h = game->GetRenderer()->getWindowHeigh();
+	float w = (float)game->GetRenderer()->getWindowWidth();
+	float h = (float)game->GetRenderer()->getWindowHeigh();
 
 	m_mdl = glm::mat4(1.0f);
 	//Ugliest way to proceed, meh, at least it works well
 	m_mdl = glm::translate(m_mdl, glm::vec3(posx + (sizex / 2), posy + (sizey / 2), 0.0f));
 	m_mdl = glm::scale(m_mdl, glm::vec3(sizex / 2, sizey / 2, 0.0f));
 
-	m_proj = glm::ortho(0.0f, 1280.0f, (float)1280.0f / (w / h), 0.0f, -1.0f, 1.0f);
+	m_proj = glm::ortho(0.0f, (float)1280, (float)1280 / (w / h), 0.0f, -1.0f, 1.0f);
 
 	m_surface = std::make_unique<RD_Quad>();
 	m_surface->Bufferize();
